Fix docs story mocks for Chromatic
Register story-local story-docs services without Vitest mocks so internal Storybook builds can index and publish docs block stories under the docgen server flag.
