+++ code/addons/docs/src/manifest.test.ts
@@ -70,7 +70,7 @@ describe('experimental_manifests', () => {
             id: 'example-component',
             path: './Example.stories.tsx',
             name: 'Example',
-            stories: {},
+            stories: [],
             jsDocTags: {},
           },
         },
@@ -104,7 +104,7 @@ describe('experimental_manifests', () => {
             id: 'example',
             path: './Example.stories.tsx',
             name: 'Example',
-            stories: {},
+            stories: [],
             jsDocTags: {},
           },
         },
@@ -140,7 +140,7 @@ describe('experimental_manifests', () => {
             id: 'example',
             path: './Example.stories.tsx',
             name: 'Example',
-            stories: {},
+            stories: [],
             jsDocTags: {},
           },
         },
@@ -213,7 +213,7 @@ describe('experimental_manifests', () => {
             id: 'example',
             path: './Example.stories.tsx',
             name: 'Example',
-            stories: {},
+            stories: [],
             jsDocTags: {},
           },
         },
@@ -275,7 +275,7 @@ describe('experimental_manifests', () => {
             id: 'example',
             path: './Example.stories.tsx',
             name: 'Example',
-            stories: {},
+            stories: [],
             jsDocTags: {},
           },
         },
@@ -402,7 +402,7 @@ describe('experimental_manifests', () => {
             id: 'example',
             path: './Example.stories.tsx',
             name: 'Example',
-            stories: {},
+            stories: [],
             jsDocTags: {},
           },
         },
